Feature/igand 673/igand 676 vt display titles list

+++ app/src/main/java/igraal/com/poc_deezer_vincent/manager/PlaylistManager.java
@@ -31,7 +31,7 @@ public class PlaylistManager {
         }
     }
 
-    public Observable<RealmPlaylist> getPlaylistById(int playlistId) {
+    public Observable<RealmPlaylist> getPlaylistById(long playlistId) {
         Observable<PlaylistServiceResponse> playlistServiceResponseObservable = service.getPlaylist(playlistId);
         return playlistServiceResponseObservable.flatMap(playlistServiceResponse ->
                 Observable.just(new RealmPlaylist(playlistServiceResponse)));
