+++ app/src/main/java/igraal/com/poc_deezer_vincent/object/jsonobject/PlaylistServiceResponse.java
@@ -9,7 +9,7 @@ public class PlaylistServiceResponse {
     TracksJson tracks;
     CreatorJson creator;
 
-    int id;
+    long id;
     String title;
     String description;
     int nb_tracks;
@@ -35,11 +35,11 @@ public class PlaylistServiceResponse {
         this.creator = creator;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -65,6 +65,7 @@ public class PlaylistServiceResponse {
 
     public void setNb_tracks(int nb_tracks) {
         this.nb_tracks = nb_tracks;
+
     }
 
     public String getPicture() {
